Renamed Namespace of Windows.Overlay to Windows.PriceCheck
