+++ src/Sidekick/App.xaml
@@ -3,7 +3,7 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:adonisUi="clr-namespace:AdonisUI;assembly=AdonisUI"
-    xmlns:converters="clr-namespace:Sidekick.Windows.Overlay.Converters"
+    xmlns:converters="clr-namespace:Sidekick.Windows.PriceCheck.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     ShutdownMode="OnExplicitShutdown"
