+++ src/Sidekick/Windows/PriceCheck/UserControls/ItemListingControl.xaml
@@ -1,5 +1,5 @@
 <UserControl
-    x:Class="Sidekick.Windows.Overlay.UserControls.ItemListingControl"
+    x:Class="Sidekick.Windows.PriceCheck.UserControls.ItemListingControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
